DATABASE-38 (#43)
* flyway tables creation abd init

* flyway's scripts names fix

* flyway's scripts names fix

* spring jpa (https://github.com/we-are-comrades/billing/issues/38)

* fix long to Long

* delete sql script V3__TABLES_CREATION...

* fix long to Long

* test for Account hashCode

* fix toString & test

+++ src/main/java/com/vectree/billing/service/UserService.java
@@ -13,9 +13,9 @@ public interface UserService {
 
     void save(User user);
 
-    void delete(int id);
+    void delete(Long id);
 
-    User getOne(int id);
+    User getOne(Long id);
 
     User findByUsername(String username);
